use decompress object with zlib
There are cases when zlib.compress returns 
Z_BUF_ERROR (-5) for string input. This is not an
error as written in the zlib manual. A simple
workaround is to use decompress objects. Note
that the Z_BUF_ERROR exception thrown by the
python zlib module for decompress might be actually
a bug in the module.